removed launchpad and relocated files
