+++ src/pages/NFTs/Collection/AggModals/BidNFTModal.tsx
@@ -17,7 +17,6 @@ import { GFX_LINK } from '../../../../styles'
 import { checkMobile, formatSOLDisplay, formatSOLNumber, notify, truncateAddress } from '../../../../utils'
 import { AppraisalValueSmall } from '../../../../utils/GenericDegsin'
 import tw from 'twin.macro'
-import styled from 'styled-components'
 import 'styled-components/macro'
 import {
   AH_FEE_ACCT,
