+++ src/pages/NFTs/Profile/ProfilePageSidebar.tsx
@@ -6,7 +6,7 @@ import { PopupProfile } from './PopupProfile'
 import { Share } from '../Share'
 import { generateTinyURL } from '../../../api/tinyUrl'
 import { WRAPPED_SOL_MINT } from '@jup-ag/core'
-import { formatNumber } from '../launchpad/candyMachine/utils'
+import { formatNumber } from '../../../web3/utils_launchpad'
 import { IAppParams } from '../../../types/app_params'
 import styled from 'styled-components'
 import tw from 'twin.macro'
