--- src/pages/NFTs/launchpad/customSC/config.ts
@@ -1,28 +1,0 @@
-import { PublicKey } from '@solana/web3.js'
-
-export const pdaWhitelistSeed = 'whitelist-config'
-
-export const MAGIC_HAT_PROGRAM_V2_ID = new PublicKey('G317XThznKMemjmSBEekMSsPedT9krKcAeRmE39drhNA')
-export const MAGIC_HAT_CREATOR = new PublicKey('4bDer2d1P357AhB8YN13b1XXBAh89d2xHex9qrM3xEtt')
-export const MAGIC_HAT_ID = new PublicKey('HvVtTCvHwuwz8YkmgindHHuXCkRAmZTR1ey5MG1v6293')
-
-export const COMMUNITY_TIME = 1658751368
-export const PUBLIC_TIME = 1658751368
-
-// export const PUBLIC_5_TIME = 1658862000
-// export const TEAM_TIME = 1658844000
-// export const OG_TIME = 1658851200
-// export const WHITELIST_TIME = 1658854800
-export const pdaSeed = 'wallet-whitelist'
-
-export const whiteListAddresses = ['9tPEcwyqwNsrgrJHrqe6tMZYD6cx6fgSv6xA297XQsDD']
-
-// export const whiteListAddressesFour = ['CqUXdhQHQuJVdch5ovonSmhpB5VYh1G5t7Wh415Pu1Ec']
-
-export const whiteListAddressesOne = ['3QiJodq6JbmhmzZ7Rf7LbtsExj2m5eMFndVw4iEPW8q2']
-
-// export const whiteListAddressesThree = ['3fTcnXKoWMoczbThdo2vxc19VMJiZ6KBDqdE55XKb8Kh']
-
-// export const priceWhitelist = 0.01
-//
-// export const pricePublic = 0.02
