+++ src/web3/connection.ts
@@ -8,18 +8,12 @@ import {
   SimulatedTransactionResponse,
   Transaction,
   TransactionInstruction,
-  TransactionSignature,
-  SystemProgram,
-  NONCE_ACCOUNT_LENGTH,
-  NonceAccount,
-  PublicKey
+  TransactionSignature
 } from '@solana/web3.js'
 import { WalletNotConnectedError } from '@solana/wallet-adapter-base'
-import { web3 } from '@project-serum/anchor'
 import { WalletContextState } from '@solana/wallet-adapter-react'
-import { sendNonceTransaction } from '../../../../api/NFTLaunchpad'
-import { perpsNotify } from '../../../../utils/perpsNotifications'
-import { confirmTransaction } from '../../../../web3'
+import { perpsNotify } from '../utils/perpsNotifications'
+import { confirmTransaction } from './index'
 
 interface BlockhashAndFeeCalculator {
   blockhash: string
@@ -53,173 +47,6 @@ export const getErrorForTransaction = async (connection: Connection, txid: strin
   }
 
   return errors
-}
-
-export enum SequenceType {
-  Sequential,
-  Parallel,
-  StopOnFailure
-}
-
-export async function sendTransactionsWithManualRetry(
-  connection: Connection,
-  wallet: WalletContextState,
-  ixs: TransactionInstruction[][],
-  signers: Keypair[][]
-): Promise<(string | undefined)[]> {
-  let instructions = ixs
-  let stopPoint = 0
-  let tries = 0
-  let lastInstructionsLength = null
-  const toRemoveSigners: Record<number, boolean> = {}
-  instructions = instructions.filter((instr, i) => {
-    if (instr.length > 0) {
-      return true
-    } else {
-      toRemoveSigners[i] = true
-      return false
-    }
-  })
-  let ids: string[] = []
-  let filteredSigners = signers.filter((_, i) => !toRemoveSigners[i])
-
-  while (stopPoint < instructions.length && tries < 3) {
-    instructions = instructions.slice(stopPoint, instructions.length)
-    filteredSigners = filteredSigners.slice(stopPoint, filteredSigners.length)
-
-    if (instructions.length === lastInstructionsLength) tries = tries + 1
-    else tries = 0
-
-    try {
-      if (instructions.length === 1) {
-        const id = await sendTransactionWithRetry(
-          connection,
-          wallet,
-          instructions[0],
-          filteredSigners[0],
-          'single'
-        )
-        ids.push(id.txid)
-        stopPoint = 1
-      } else {
-        const { txs } = await sendTransactions(
-          connection,
-          wallet,
-          instructions,
-          filteredSigners,
-          SequenceType.StopOnFailure,
-          'single'
-        )
-        ids = ids.concat(txs.map((t) => t.txid))
-      }
-    } catch (e) {
-      console.error(e)
-    }
-    console.log(
-      'Died on ',
-      stopPoint,
-      'retrying from instruction',
-      instructions[stopPoint],
-      'instructions length is',
-      instructions.length
-    )
-    lastInstructionsLength = instructions.length
-  }
-
-  return ids
-}
-
-export const sendTransactions = async (
-  connection: Connection,
-  wallet: WalletContextState,
-  instructionSet: TransactionInstruction[][],
-  signersSet: Keypair[][],
-  sequenceType: SequenceType = SequenceType.Parallel,
-  commitment: Commitment = 'singleGossip',
-  successCallback: (txid?: string, ind?: number) => void = () => {}, //eslint-disable-line
-  failCallback: (reason?: Transaction, ind?: number) => boolean = () => false, //eslint-disable-line
-  block?: BlockhashAndFeeCalculator,
-  beforeTransactions: Transaction[] = [],
-  afterTransactions: Transaction[] = []
-): Promise<{ number: number; txs: { txid: string; slot: number }[] }> => {
-  if (!wallet.publicKey) throw new WalletNotConnectedError()
-
-  const unsignedTxns: Transaction[] = beforeTransactions
-
-  if (!block) {
-    //eslint-disable-next-line
-    block = await connection.getLatestBlockhash(commitment)
-  }
-
-  for (let i = 0; i < instructionSet.length; i++) {
-    const instructions = instructionSet[i]
-    const signers = signersSet[i]
-
-    if (instructions.length === 0) {
-      continue
-    }
-
-    const transaction = new Transaction()
-    instructions.forEach((instruction) => transaction.add(instruction))
-    transaction.recentBlockhash = block.blockhash
-    transaction.setSigners(
-      // fee payed by the wallet owner
-      wallet.publicKey,
-      ...signers.map((s) => s.publicKey)
-    )
-
-    if (signers.length > 0) {
-      transaction.partialSign(...signers)
-    }
-
-    unsignedTxns.push(transaction)
-  }
-  unsignedTxns.push(...afterTransactions)
-
-  const partiallySignedTransactions = unsignedTxns.filter((t) =>
-    t.signatures.find((sig) => sig.publicKey.equals(wallet.publicKey))
-  )
-  const fullySignedTransactions = unsignedTxns.filter(
-    (t) => !t.signatures.find((sig) => sig.publicKey.equals(wallet.publicKey))
-  )
-  let signedTxns = await wallet.signAllTransactions(partiallySignedTransactions)
-  signedTxns = fullySignedTransactions.concat(signedTxns)
-  const pendingTxns: Promise<{ txid: string; slot: number }>[] = []
-
-  console.log('Signed txns length', signedTxns.length, 'vs handed in length', instructionSet.length)
-  for (let i = 0; i < signedTxns.length; i++) {
-    const signedTxnPromise = sendSignedTransaction({
-      connection,
-      signedTransaction: signedTxns[i]
-    })
-
-    if (sequenceType !== SequenceType.Parallel) {
-      try {
-        await signedTxnPromise.then(({ txid }) => successCallback(txid, i))
-        pendingTxns.push(signedTxnPromise)
-      } catch (e) {
-        console.log('Failed at txn index:', i)
-        console.log('Caught failure:', e)
-
-        failCallback(signedTxns[i], i)
-        if (sequenceType === SequenceType.StopOnFailure) {
-          return {
-            number: i,
-            txs: await Promise.all(pendingTxns)
-          }
-        }
-      }
-    } else {
-      pendingTxns.push(signedTxnPromise)
-    }
-  }
-
-  if (sequenceType !== SequenceType.Parallel) {
-    const result = await Promise.all(pendingTxns)
-    return { number: signedTxns.length, txs: result }
-  }
-
-  return { number: signedTxns.length, txs: await Promise.all(pendingTxns) }
 }
 
 export const sendTransaction = async (
@@ -549,126 +376,6 @@ export const sendPerpsTransactions = async (
   return ixResponse
 }
 
-export const sendTransactionWithRetry = async (
-  connection: Connection,
-  wallet: WalletContextState,
-  instructions: TransactionInstruction[],
-  signers: Keypair[],
-  commitment: Commitment = 'singleGossip',
-  includesFeePayer = false,
-  block?: BlockhashAndFeeCalculator,
-  beforeSend?: () => void
-): Promise<{ txid: string; slot: number }> => {
-  if (!wallet.publicKey) throw new WalletNotConnectedError()
-
-  let transaction = new Transaction()
-  instructions.forEach((instruction) => transaction.add(instruction))
-  transaction.recentBlockhash = (block || (await connection.getLatestBlockhash(commitment))).blockhash
-
-  if (includesFeePayer) {
-    transaction.setSigners(...signers.map((s) => s.publicKey))
-  } else {
-    transaction.setSigners(
-      // fee payed by the wallet owner
-      wallet.publicKey,
-      ...signers.map((s) => s.publicKey)
-    )
-  }
-
-  if (signers.length > 0) {
-    transaction.partialSign(...signers)
-  }
-  if (!includesFeePayer) {
-    transaction = await wallet.signTransaction(transaction)
-  }
-
-  if (beforeSend) {
-    beforeSend()
-  }
-
-  const { txid, slot } = await sendSignedTransaction({
-    connection,
-    signedTransaction: transaction
-  })
-
-  return { txid, slot }
-}
-
-export const getUnixTs = (): number => new Date().getTime() / 1000
-
-export async function sendSignedTransaction({
-  signedTransaction,
-  connection,
-  timeout = DEFAULT_TIMEOUT
-}: {
-  signedTransaction: Transaction
-  connection: Connection
-  sendingMessage?: string
-  sentMessage?: string
-  successMessage?: string
-  timeout?: number
-}): Promise<{ txid: string; slot: number }> {
-  const rawTransaction = signedTransaction.serialize()
-
-  const startTime = getUnixTs()
-  let slot = 0
-  const txid: TransactionSignature = await connection.sendRawTransaction(rawTransaction, {
-    skipPreflight: true
-  })
-
-  console.log('Started awaiting confirmation for', txid)
-
-  let done = false
-  ;(async () => {
-    while (!done && getUnixTs() - startTime < timeout) {
-      connection.sendRawTransaction(rawTransaction, {
-        skipPreflight: true
-      })
-      await sleep(500)
-    }
-  })()
-  try {
-    console.log('before confirmation')
-    const confirmation = await awaitTransactionSignatureConfirmation(txid, timeout, connection, 'recent', true)
-    console.log('confirmation is ', confirmation)
-    if (!confirmation) throw new Error('Timed out awaiting confirmation on transaction')
-
-    if (confirmation.err) {
-      console.error(confirmation.err)
-      throw new Error('Transaction failed: Custom instruction error')
-    }
-
-    slot = confirmation?.slot || 0
-  } catch (err: any) {
-    console.error('Timeout Error caught', err)
-    if (err.timeout) {
-      throw new Error('Timed out awaiting confirmation on transaction')
-    }
-    let simulateResult: SimulatedTransactionResponse | null = null
-    try {
-      simulateResult = (await simulateTransaction(connection, signedTransaction, 'single')).value
-    } catch (e) {
-      console.log(e)
-    }
-    if (simulateResult && simulateResult.err) {
-      if (simulateResult.logs) {
-        for (let i = simulateResult.logs.length - 1; i >= 0; --i) {
-          const line = simulateResult.logs[i]
-          if (line.startsWith('Program log: ')) {
-            throw new Error('Transaction failed: ' + line.slice('Program log: '.length))
-          }
-        }
-      }
-      throw new Error(JSON.stringify(simulateResult.err))
-    }
-  } finally {
-    done = true
-  }
-
-  console.log('Latency', txid, getUnixTs() - startTime)
-  return { txid, slot }
-}
-
 async function simulateTransaction(
   connection: Connection,
   transaction: Transaction,
@@ -802,99 +509,4 @@ async function awaitTransactionSignatureConfirmation(
 }
 export function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms))
-}
-
-export async function nonceInstructions(
-  connection: Connection,
-  feePayer: { publicKey: PublicKey }
-): Promise<void> {
-  const nonceAccount = web3.Keypair.generate()
-  const instrctions = []
-
-  instrctions.push(
-    SystemProgram.createAccount({
-      fromPubkey: feePayer.publicKey,
-      newAccountPubkey: nonceAccount.publicKey,
-      lamports: await connection.getMinimumBalanceForRentExemption(NONCE_ACCOUNT_LENGTH),
-      space: NONCE_ACCOUNT_LENGTH,
-      programId: SystemProgram.programId
-    })
-  )
-
-  instrctions.push(
-    SystemProgram.nonceInitialize({
-      noncePubkey: nonceAccount.publicKey, // nonce account pubkey
-      authorizedPubkey: feePayer.publicKey // nonce account authority (for advance and close)
-    })
-  )
-
-  instrctions.push(
-    SystemProgram.nonceAdvance({
-      noncePubkey: nonceAccount.publicKey,
-      authorizedPubkey: feePayer.publicKey
-    })
-  )
-}
-
-export const sendTransactionsNonce = async (
-  connection: Connection,
-  wallet: WalletContextState,
-  instructionSet: TransactionInstruction[][],
-  signersSet: Keypair[][],
-  beforeTransactions: Transaction[] = [],
-  afterTransactions: Transaction[] = [],
-  nonce: PublicKey,
-  collectionId: string,
-  walletAddress: string
-): Promise<{ number: number; txs: { txid: string; slot: number }[] }> => {
-  if (!wallet.publicKey) throw new WalletNotConnectedError()
-
-  const unsignedTxns: Transaction[] = beforeTransactions
-
-  for (let i = 0; i < instructionSet.length; i++) {
-    const instructions = instructionSet[i]
-    const signers = signersSet[i]
-
-    if (instructions.length === 0) {
-      continue
-    }
-
-    const transaction = new Transaction()
-    instructions.forEach((instruction) => transaction.add(instruction))
-    const accountInfo = await connection.getAccountInfo(nonce)
-    const nonceAccount = NonceAccount.fromAccountData(accountInfo.data)
-
-    transaction.recentBlockhash = nonceAccount.nonce
-
-    transaction.setSigners(
-      // fee payed by the wallet owner
-      wallet.publicKey,
-      ...signers.map((s) => s.publicKey)
-    )
-
-    if (signers.length > 0) {
-      transaction.partialSign(...signers)
-    }
-
-    unsignedTxns.push(transaction)
-  }
-  unsignedTxns.push(...afterTransactions)
-
-  const partiallySignedTransactions = unsignedTxns.filter((t) =>
-    t.signatures.find((sig) => sig.publicKey.equals(wallet.publicKey))
-  )
-  const fullySignedTransactions = unsignedTxns.filter(
-    (t) => !t.signatures.find((sig) => sig.publicKey.equals(wallet.publicKey))
-  )
-  let signedTxns = await wallet.signAllTransactions(partiallySignedTransactions)
-  signedTxns = fullySignedTransactions.concat(signedTxns)
-
-  console.log('Signed txns length', signedTxns.length, 'vs handed in length', instructionSet.length)
-  try {
-    const response = await sendNonceTransaction(signedTxns[0].serialize(), collectionId, walletAddress)
-    return response.data
-  } catch (e) {
-    console.log(e)
-    return null
-  }
 }
